+++ pkg/services/auth_test.go
@@ -1,0 +1,40 @@
+package services
+
+import (
+	"context"
+
+	"github.com/cmsgov/easi-app/pkg/appcontext"
+	"github.com/cmsgov/easi-app/pkg/authn"
+)
+
+func (s ServicesTestSuite) TestAuthorizeRequireGRTJobCode() {
+	fnAuth := NewAuthorizeRequireGRTJobCode()
+	nonGRT := authn.EUAPrincipal{EUAID: "FAKE", JobCodeEASi: true, JobCodeGRT: false}
+	yesGRT := authn.EUAPrincipal{EUAID: "FAKE", JobCodeEASi: true, JobCodeGRT: true}
+
+	testCases := map[string]struct {
+		ctx     context.Context
+		allowed bool
+	}{
+		"anonymous": {
+			ctx:     context.Background(),
+			allowed: false,
+		},
+		"non grt": {
+			ctx:     appcontext.WithPrincipal(context.Background(), &nonGRT),
+			allowed: false,
+		},
+		"has grt": {
+			ctx:     appcontext.WithPrincipal(context.Background(), &yesGRT),
+			allowed: true,
+		},
+	}
+
+	for name, tc := range testCases {
+		s.Run(name, func() {
+			ok, err := fnAuth(tc.ctx)
+			s.NoError(err)
+			s.Equal(tc.allowed, ok)
+		})
+	}
+}
